+++ index.js
@@ -4,22 +4,30 @@
 var _options = {
   svg: {
     xmlns: 'http://www.w3.org/2000/svg',
+    'xmlns:xlink': 'http://www.w3.org/1999/xlink',
     style: 'position:absolute; width: 0; height: 0'
   },
   loop: 2,
   svgoOptions: {},
   name: 'sprite.[hash].svg',
-  prefix: 'icon-'
+  prefix: 'icon-',
+  template: __dirname + '/templates/layout.jade'
 };
 
 // Depends
 var _ = require('lodash');
 var path = require('path');
-var glob = require('glob');
+var url = require('url');
 var slash = require('slash');
 var utils = require('./helpers/utils');
-var ConcatSource = require('webpack/lib/ConcatSource');
 var ModuleFilenameHelpers = require('webpack/lib/ModuleFilenameHelpers');
+
+var ConcatSource;
+try {
+  ConcatSource = require('webpack/lib/ConcatSource');       // webpack 1.x
+} catch (e) {
+  ConcatSource = require('webpack-sources').ConcatSource;   // webpack 2.x
+}
 
 /**
  * Constructor
@@ -30,38 +38,11 @@ var ModuleFilenameHelpers = require('webpack/lib/ModuleFilenameHelpers');
  */
 var WebpackSvgStore = function(input, output, options) {
   // set attributes
-  this.input    = input;
-  this.output   = output;
-  this.options  = _.merge({}, _options, options);
+  this.input   = input;
+  this.output  = output;
+  this.options = _.merge({}, _options, options);
 
   return this;
-};
-
-/**
- * Build files map
- * @param  {string} input Destination path
- * @return {array}        Array of paths
- */
-WebpackSvgStore.prototype.filesMap = function(input, cb) {
-  var files = [];
-  var data = input;
-  // in case if array was passed
-  if (data instanceof Array) {
-    data.forEach(function(source) {
-      this.filesMap(source, function(fileList) {
-        files = files.concat(fileList);
-      });
-    });
-    cb(files);
-  } else {
-    glob(data, {nodir: true}, function(error, fileList) {
-      if (error) {
-        throw error;
-      }
-      // slice off pattern
-      cb(fileList);
-    });
-  }
 };
 
 /**
@@ -70,54 +51,73 @@ WebpackSvgStore.prototype.filesMap = function(input, cb) {
  * @return {[type]}          [description]
  */
 WebpackSvgStore.prototype.apply = function(compiler) {
-  var chunkWrapper;
   var publicPath;
 
-  var self = this;
   var options = this.options;
+
   var inputFolder = this.input;
   var outputFolder = this.output;
   var spriteName = this.options.name;
-
-  // prepare input / output folders
-  utils.prepareFolder(inputFolder);
-  utils.prepareFolder(outputFolder);
+  var lastXhrText;
 
   // subscribe to webpack emit state
-  compiler.plugin('compilation', function(compilation) {
-    publicPath = compilation.getStats().toJson().publicPath || '/';
-    self.filesMap(inputFolder, function(files) {
-      var fileContent = utils.createSprite(utils.parseFiles(files, options));
-      var fileName = utils.hash(fileContent, spriteName);
-      var filePath = path.join(outputFolder, fileName);
+  compiler.plugin('emit', function(compilation, callback) {
+    // path into dist absolute path
+    publicPath = compilation.getStats().toJson().publicPath;
 
-      // fallback for windows backslashes
-      var fullPath = slash(path.join(publicPath, filePath));
+    if (!publicPath) {
+      publicPath = path.isAbsolute(outputFolder)
+        ? outputFolder
+        : '/'
+      ;
+    }
 
-      compilation.assets[filePath] = {
+    // prepare output folder
+    utils.prepareFolder(outputFolder);
+
+    utils.filesMap(inputFolder, function(files) {
+      var fullPath;
+      var fileContent;
+      var fileName;
+      var filePath;
+
+      if (!utils.filesChanged(files)) {
+        injectXhr(options, compilation, lastXhrText);
+        return callback();
+      }
+
+      fileContent = utils.createSprite(utils.parseFiles(files, options), options.template);
+      fileName = utils.hash(fileContent, spriteName);
+      filePath = path.join(outputFolder, fileName);
+
+      // resolve with node url
+      fullPath = url.resolve(publicPath, filePath);
+
+      compilation.assets[slash(filePath)] = {
         size: function() { return Buffer.byteLength(fileContent, 'utf8'); },
         source: function() { return new Buffer(fileContent); }
       };
 
-      // if chunk enable
-      if (options && options.chunk) {
-        chunkWrapper = options.chunk;
-        compilation.plugin('optimize-chunk-assets', function(chunks, callback) {
-          chunks.forEach(function(chunk) {
-            if (options.entryOnly && !chunk.initial) return;
-            if (chunk.name === chunkWrapper) {
-              chunk.files.filter(ModuleFilenameHelpers.matchObject.bind(undefined, options)).forEach(function(file) {
-                if (/\.js?$/.test(file)) {
-                  compilation.assets[file] = new ConcatSource(utils.svgXHR(fullPath, options.baseUrl), '\n', compilation.assets[file]);
-                }
-              });
-            }
-          });
-          callback();
-        });
-      }
+      lastXhrText = utils.svgXHR(fullPath, options.baseUrl);
+      injectXhr(options, compilation, lastXhrText);
+
+      callback();
     });
   });
+
+  function injectXhr(opt, compilation, lxt) {
+    var chunk;
+    // if chunk enable apply to chunk
+    if (opt && opt.chunk) {
+      chunk = _.find(compilation.chunks, {name: opt.chunk});
+      if (!chunk) return;
+      chunk.files.filter(ModuleFilenameHelpers.matchObject.bind(undefined, opt)).forEach(function(file) {
+        if (/\.js?$/.test(file)) {
+          compilation.assets[file] = new ConcatSource(lxt, '\n', compilation.assets[file]);
+        }
+      });
+    }
+  }
 };
 
 /**