+++ CHANGELOG.md
@@ -1,4 +1,27 @@
+### 2.2.2
+
+- Add default`xmlns:xlink` attribute to result svg
+
+### 2.2.1
+
+- Use new ConcatSource dependency for webpack 2.x
+
+### 2.2.0
+
+- Add `template` option for custom jade layout
+- Add file caching to fix multiple sprite creation
+- Use `emit` instead `compilation` state for comfortable work in devmode
+
+### 2.1.5
+
+- Change `glob` with `globby` to support multiple patterns
+
+### 2.0.3 - 2.1.4
+
+- Сhanges during this period have been lost and forgotten :'(
+
 ### 2.0.2 Global refactoring
+
 - Move deprecated cheerio parsing to hell
 - Performance optimization
 - Remove some magic initial params from plugin init config